Factor out messageType.
