+++ apps/frontend/src/lib/langgraph/types.ts
@@ -1,7 +1,22 @@
-import type { AIMessage, ToolMessage } from '@langchain/langgraph-sdk';
-
-export interface FixedAIMessage extends Omit<AIMessage, 'type'> {
-	type: 'AIMessageChunk';
+export interface BaseMessage {
+	type: string;
+	text: string;
+	id: string;
 }
 
-export type FixedMessage = FixedAIMessage | ToolMessage;
+export interface AIMessage extends BaseMessage {
+	type: 'ai';
+}
+
+export interface UserMessage extends BaseMessage {
+	type: 'user';
+}
+
+export interface ToolMessage extends BaseMessage {
+	type: 'tool';
+	tool_name: string;
+	payload?: Record<string, unknown>;
+	status?: 'success' | 'error';
+}
+
+export type Message = AIMessage | UserMessage | ToolMessage;
