--- apps/frontend/src/lib/types/messageTypes.ts
@@ -1,28 +1,0 @@
-export interface BaseMessage {
-	type: string;
-	text: string;
-	id: string;
-}
-
-export interface AIMessage extends BaseMessage {
-	type: 'ai';
-}
-
-export interface UserMessage extends BaseMessage {
-	type: 'user';
-}
-
-export interface ToolMessage extends BaseMessage {
-	type: 'tool';
-	tool_name: string;
-	payload?: Record<string, unknown>;
-	status?: 'success' | 'error';
-}
-
-export type Message = AIMessage | UserMessage | ToolMessage;
-
-export interface ChatSuggestion {
-	title: string;
-	description: string;
-	suggestedText: string;
-}
